move EdfSignal to a separate module, same for the related tests
