+++ edfio/__init__.py
@@ -2,11 +2,11 @@ from edfio.edf import (
     AnonymizedDateError,
     Edf,
     EdfAnnotation,
-    EdfSignal,
     Patient,
     Recording,
     read_edf,
 )
+from edfio.edf_signal import EdfSignal
 
 __version__ = "0.0.0"
