+++ edfio/_utils.py
@@ -2,18 +2,7 @@ from __future__ import annotations
 
 import datetime
 import inspect
-from typing import Any, Callable, NamedTuple
-
-
-def calculate_gain_and_offset(
-    digital_min: int,
-    digital_max: int,
-    physical_min: float,
-    physical_max: float,
-) -> tuple[float, float]:
-    gain = (physical_max - physical_min) / (digital_max - digital_min)
-    offset = physical_max / gain - digital_max
-    return gain, offset
+from typing import Any
 
 
 def repr_from_init(obj: Any) -> str:
@@ -21,16 +10,6 @@ def repr_from_init(obj: Any) -> str:
     for name in inspect.signature(obj.__class__).parameters:
         parameters.append(f"{name}={getattr(obj, name)!r}")
     return f"{obj.__class__.__name__}({', '.join(parameters)})"
-
-
-class IntRange(NamedTuple):
-    min: int
-    max: int
-
-
-class FloatRange(NamedTuple):
-    min: float
-    max: float
 
 
 _MONTH_NAMES = (
@@ -76,20 +55,6 @@ def encode_annotation_duration(duration: float) -> str:
     if string[-1] == ".":
         return string[:-1]
     return string
-
-
-def round_float_to_8_characters(
-    value: float,
-    round_func: Callable[[float], int],
-) -> float:
-    if isinstance(value, int) or value.is_integer():
-        return value
-    length = 8
-    integer_part_length = str(value).find(".")
-    if integer_part_length == length:
-        return round_func(value)
-    factor = 10 ** (length - 1 - integer_part_length)
-    return round_func(value * factor) / factor
 
 
 def validate_subfields(subfields: dict[str, str]) -> None:
