+++ tests/test_edf.py
@@ -3,7 +3,6 @@ from __future__ import annotations
 import datetime
 import io
 import json
-import re
 import tempfile
 from pathlib import Path
 from typing import Literal
@@ -20,8 +19,8 @@ from edfio import (
     Recording,
     read_edf,
 )
-from edfio._utils import FloatRange, IntRange
 from edfio.edf import _calculate_num_data_records, _create_annotations_signal
+from edfio.edf_signal import _FloatRange
 from tests import TEST_DATA_DIR
 
 EDF_FILE = TEST_DATA_DIR / "short_psg.edf"
@@ -68,112 +67,6 @@ def test_read_edf():
 def sine(duration, f, fs):
     t = np.arange(duration * fs) / fs
     return np.sin(2 * np.pi * f * t)
-
-
-@pytest.fixture()
-def dummy_edf_signal() -> EdfSignal:
-    data = sine(5, 2, 128)
-    return EdfSignal(data, 128)
-
-
-def test_edf_signal_init_minimal():
-    data = sine(5, 2, 128)
-    sig = EdfSignal(data, 128)
-    tolerance = np.ptp(data) * 2**-16
-    np.testing.assert_allclose(
-        sig.data,
-        data,
-        atol=tolerance,
-    )
-    assert sig.sampling_frequency == 128
-    assert sig.physical_min == data.min()
-    assert sig.physical_max == data.max()
-
-
-def test_edf_signal_init_all_parameters():
-    data = sine(5, 2, 128)
-    physical_range = FloatRange(-500, 500)
-    digital_range = IntRange(-2048, 2047)
-    params = {
-        "sampling_frequency": 128,
-        "label": "EEG Fpz-Cz",
-        "transducer_type": "AgAgCl electrode",
-        "physical_dimension": "uV",
-        "physical_range": physical_range,
-        "digital_range": digital_range,
-        "prefiltering": "HP:0.1Hz LP:75Hz",
-    }
-    sig = EdfSignal(data, **params)
-    tolerance = (physical_range.max - physical_range.min) / (
-        digital_range.max - digital_range.min
-    )
-    np.testing.assert_allclose(
-        sig.data,
-        data,
-        atol=tolerance,
-    )
-    for name, value in params.items():
-        assert getattr(sig, name) == value
-    assert sig.reserved == ""
-
-
-@pytest.mark.parametrize(
-    "field_name",
-    [
-        "samples_per_data_record",
-        "reserved",
-    ],
-)
-def test_edf_signal_field_cannot_be_set_publicly(field_name: str):
-    signal = EdfSignal(np.arange(10), 1)
-    with pytest.raises(AttributeError, match="can't set attribute"):
-        setattr(signal, field_name, None)
-
-
-EDFSIGNAL_SETTER_TEST_FIELDS = [
-    ("_label", 16, "EEG FPz-Cz"),
-    ("transducer_type", 80, "AgAgCl electrode"),
-    ("physical_dimension", 8, "uV"),
-    ("prefiltering", 80, "HP:0.1Hz LP:75Hz"),
-]
-
-
-@pytest.mark.parametrize(
-    ("field_name", "field_length", "value"),
-    EDFSIGNAL_SETTER_TEST_FIELDS,
-)
-def test_edfsignal_setter_raises_error_if_field_length_is_exceeded(
-    dummy_edf_signal: EdfSignal,
-    field_name: str,
-    field_length: int,
-    value: str,
-):
-    prev_val = getattr(dummy_edf_signal, field_name)
-    with pytest.raises(ValueError, match="exceeds maximum field length"):
-        setattr(dummy_edf_signal, field_name, "X" * (field_length + 1))
-    assert getattr(dummy_edf_signal, field_name) == prev_val
-
-
-@pytest.mark.parametrize(
-    ("field_name", "field_length", "value"),
-    EDFSIGNAL_SETTER_TEST_FIELDS,
-)
-def test_edfsignal_setter_sets_raw_data_fields(
-    dummy_edf_signal: EdfSignal,
-    field_name: str,
-    field_length: int,
-    value: str,
-):
-    setattr(dummy_edf_signal, field_name, value)
-    assert getattr(dummy_edf_signal, field_name) == value
-    assert getattr(dummy_edf_signal, "_" + field_name) == str(value).ljust(
-        field_length
-    ).encode("ascii")
-
-
-def test_edf_signal_samples_per_data_record_not_set(dummy_edf_signal: EdfSignal):
-    with pytest.raises(AttributeError, match="no attribute '_samples_per_data_record'"):
-        dummy_edf_signal.samples_per_data_record
 
 
 EDF_IDENTIFICATION_SETTER_TEST_FIELDS = [
@@ -249,19 +142,6 @@ def test_edf_init_with_negative_data_record_duration():
 
 
 @pytest.mark.parametrize(
-    ("signal", "expected"),
-    [
-        (EdfSignal(np.arange(5), 1), "<EdfSignal 1Hz>"),
-        (EdfSignal(np.arange(5), 1, label="ECG"), "<EdfSignal ECG 1Hz>"),
-        (EdfSignal(np.arange(5), 256.0), "<EdfSignal 256Hz>"),
-        (EdfSignal(np.arange(5), 123.456), "<EdfSignal 123.456Hz>"),
-    ],
-)
-def test_edf_signal_repr(signal: EdfSignal, expected: str):
-    assert repr(signal) == expected
-
-
-@pytest.mark.parametrize(
     ("edf", "expected"),
     [
         (
@@ -334,24 +214,6 @@ def test_data_record_duration_does_not_exactly_divide_signal_duration(
         Edf(signals, data_record_duration=data_record_duration)
 
 
-def test_edf_signal_from_raw_header_has_no_data_by_default():
-    sig = EdfSignal._from_raw_header(
-        20,
-        _label=b"".ljust(16),
-        transducer_type=b"".ljust(80),
-        physical_dimension=b"".ljust(8),
-        physical_min=b"-500".ljust(8),
-        physical_max=b"500".ljust(8),
-        digital_min=b"-2048".ljust(8),
-        digital_max=b"2047".ljust(8),
-        prefiltering=b"".ljust(80),
-        samples_per_data_record=b"1".ljust(8),
-        reserved=b"".ljust(32),
-    )
-    with pytest.raises(AttributeError, match="has no attribute '_digital'"):
-        sig.data
-
-
 def test_read_write_roundtrip(tmp_file: Path):
     edf = read_edf(EDF_FILE)
     edf.write(tmp_file)
@@ -398,37 +260,6 @@ def test_write_read_roundtrip_signal_tolerance(tmp_file: Path, signal_scale: flo
         )
 
 
-@pytest.mark.parametrize(
-    ("physical_range", "actual_data"),
-    [
-        pytest.param(
-            FloatRange(-18.1234, 112.5432),
-            [13.0, -17.9, -18.0, 112.543251],
-            id="Exceeds maximum",
-        ),
-        pytest.param(
-            FloatRange(-18.1234, 112.5432),
-            [13.0, -18.123451, -18.0, 110.0],
-            id="Exceeds minimum",
-        ),
-    ],
-)
-def test_write_data_that_exceeds_physical_range(
-    physical_range: FloatRange, actual_data: list[float], tmp_file: Path
-) -> None:
-    with pytest.raises(
-        ValueError,
-        match=re.escape(
-            f"Signal range [{min(actual_data)}, {max(actual_data)}] out of physical range: [{physical_range.min}, {physical_range.max}]"
-        ),
-    ):
-        EdfSignal(
-            data=np.asarray(actual_data, dtype="float64"),
-            sampling_frequency=1,
-            physical_range=physical_range,
-        )
-
-
 def test_physical_values_are_mapped_to_digital_values_with_minimal_error(
     tmp_file: Path,
 ):
@@ -438,7 +269,7 @@ def test_physical_values_are_mapped_to_digital_values_with_minimal_error(
         EdfSignal(
             data=np.asarray(data, dtype="float64"),
             sampling_frequency=1,
-            physical_range=FloatRange(-32768, 32767),  # same as digital range
+            physical_range=_FloatRange(-32768, 32767),  # same as digital range
         )
     ]
     Edf(signals).write(tmp_file)
@@ -495,106 +326,6 @@ def test_edf_field_cannot_be_set_publicly(field_name: str):
     edf = read_edf(EDF_FILE)
     with pytest.raises(AttributeError, match="can't set attribute|has no setter"):
         setattr(edf, field_name, None)
-
-
-def test_edf_signal_with_too_small_physical_range_fails():
-    with pytest.raises(ValueError, match="out of physical range"):
-        EdfSignal(np.arange(10), 1, physical_range=(3, 5))
-
-
-@pytest.mark.parametrize("data", [[-3, 3], [-3, 7], [-7, 7]])
-def test_edf_signal_update_data(data):
-    data = np.array(data)
-    signal = EdfSignal(np.array([-5, 5]), 1)
-    signal.update_data(data)
-    np.testing.assert_array_equal(signal.data, data)
-    np.testing.assert_array_equal(signal._digital, np.array([-32768, 32767]))
-    assert signal.physical_range == FloatRange(data.min(), data.max())
-
-
-@pytest.mark.parametrize("length", [9, 11])
-def test_edf_signal_update_data_fails_on_length_mismatch(length: int):
-    signal = EdfSignal(np.arange(10), 1)
-    with pytest.raises(ValueError, match="Signal lengths must match"):
-        signal.update_data(np.arange(length))
-
-
-def test_edf_signal_update_data_keep_physical_range():
-    signal = EdfSignal(
-        np.array([90.1, 98.3, 100.0]),
-        1,
-        physical_range=(0, 100),
-        digital_range=(0, 1000),
-    )
-    new_data = np.array([92.3, 97.5, 99.9])
-    signal.update_data(new_data, keep_physical_range=True)
-    assert signal.physical_range == (0, 100)
-    assert signal.digital_range == (0, 1000)
-    np.testing.assert_almost_equal(signal.data, new_data, decimal=14)
-    np.testing.assert_array_equal(signal._digital, new_data * 10)
-
-
-@pytest.mark.parametrize("data", [[-7, 3], [-3, 7], [-7, 7]])
-def test_edf_signal_update_data_keep_physical_range_raises_error_if_new_data_exceeds_physical_range(
-    data,
-):
-    signal = EdfSignal(np.array([-5, 5]), 1)
-    with pytest.raises(ValueError, match="Signal range .* out of physical range"):
-        signal.update_data(np.array(data), keep_physical_range=True)
-
-
-def test_edf_signal_update_data_resampling():
-    signal = EdfSignal(np.arange(10), 10, digital_range=(0, 9), physical_range=(0, 9))
-    signal.update_data(
-        np.arange(0, 10, 2), sampling_frequency=5, keep_physical_range=True
-    )
-    assert signal.sampling_frequency == 5
-    np.testing.assert_array_equal(signal.data, np.arange(0, 10, 2))
-
-
-def test_edf_signal_update_data_resampling_noninteger_sampling_rates():
-    signal = EdfSignal(
-        np.arange(11), 5.5, digital_range=(0, 10), physical_range=(0, 10)
-    )
-    signal.update_data(np.arange(0, 10), sampling_frequency=5, keep_physical_range=True)
-    assert signal.sampling_frequency == 5
-    np.testing.assert_array_equal(signal.data, np.arange(0, 10))
-
-
-def test_edf_signal_update_data_resampling_invalid_duration():
-    signal = EdfSignal(np.arange(10), 10, digital_range=(0, 9), physical_range=(0, 9))
-    with pytest.raises(ValueError, match="Signal lengths must match:"):
-        signal.update_data(np.arange(0, 10, 2.5), sampling_frequency=5)
-
-
-@pytest.mark.parametrize(
-    "sampling_frequency",
-    [-10, 0],
-)
-def test_edf_signal_update_data_resampling_non_positive_sampling_frequency(
-    sampling_frequency: float,
-):
-    signal = EdfSignal(np.arange(10), 10, digital_range=(0, 9), physical_range=(0, 9))
-    with pytest.raises(ValueError, match="Sampling frequency must be positive"):
-        signal.update_data(np.arange(0, 10, 2), sampling_frequency=sampling_frequency)
-
-
-def test_edf_signal_update_data_resampling_non_integer_samples():
-    signal = EdfSignal(np.arange(10), 10, digital_range=(0, 9), physical_range=(0, 9))
-    with pytest.raises(ValueError, match="non-integer number of samples"):
-        signal.update_data(np.arange(0, 10, 2), sampling_frequency=9.8)
-
-
-def test_edf_signal_update_data_resampling_tolerance():
-    signal = EdfSignal(np.arange(1000000), 100)
-    signal.update_data(np.arange(1000001), sampling_frequency=100.0001)
-    assert signal.sampling_frequency == 100.0001
-    assert len(signal.data) == 1000001
-
-
-def test_edf_signal_data_cannot_be_modified(dummy_edf_signal: EdfSignal):
-    with pytest.raises(ValueError, match="assignment destination is read-only"):
-        dummy_edf_signal.data[5] = -1
 
 
 def test_edf_signal_requiring_rounding_in_physical_max_is_written_and_read_correctly(
@@ -685,31 +416,10 @@ def test_write_edf_bytes_io():
     assert stream.read() == EDF_FILE.read_bytes()
 
 
-@pytest.mark.parametrize("data", [[0, np.nan, 2], [0, np.inf, 2], [0, -np.inf, 2]])
-def test_instantiating_edf_signal_with_non_finite_data_raises_error(data: list[float]):
-    with pytest.raises(ValueError, match="Signal data must contain only finite values"):
-        EdfSignal(np.array(data), 3)
-
-
 def test_write_and_read_edf_with_non_ascii_characters_in_path(tmp_path: Path):
     tmp_file = tmp_path / "带 Annotation 的 EDF 文件.edf"
     Edf([EdfSignal(np.arange(10), 1)]).write(tmp_file)
     read_edf(tmp_file)
-
-
-def test_creating_edf_signal_where_physical_min_equals_physical_max_raises_error():
-    with pytest.raises(ValueError, match="Physical minimum .* must differ from"):
-        EdfSignal(np.zeros(5), sampling_frequency=1, physical_range=(0, 0))
-
-
-def test_creating_edf_signal_where_digital_min_equals_digital_max_raises_error():
-    with pytest.raises(ValueError, match="Digital minimum .* must differ from"):
-        EdfSignal(np.arange(5), sampling_frequency=1, digital_range=(0, 0))
-
-
-def test_creating_edf_signal_with_constant_data_is_possible():
-    signal = EdfSignal(np.zeros(5), 1)
-    np.testing.assert_array_equal(signal.data, np.zeros(5))
 
 
 @pytest.mark.parametrize("data_record_duration", [0.1, 0.3, 0.5, 1.5])
@@ -1040,14 +750,6 @@ def test_writing_edfplus_works_with_data_record_duration_different_from_1(
     edf.to_bytes()
 
 
-def test_rounding_of_physical_range_does_not_produce_clipping_or_integer_overflow():
-    data = np.array([0, 0.0000014999])
-    sig = EdfSignal(data, 1)
-    np.testing.assert_allclose(sig.data, data, atol=1e-11)
-    # round((0.0000014999 - 0.0) / 0.000002 * 65535 + (-32768)) = 16380
-    assert sig._digital.tolist() == [-32768, 16380]
-
-
 @pytest.mark.parametrize(
     ("signal_duration", "data_record_duration", "result"),
     [
@@ -1262,7 +964,7 @@ def test_update_data_record_duration_method_pad_with_zero():
 
 @pytest.mark.parametrize("physical_range", [(5, 104), (-105, -6)])
 def test_update_data_record_duration_method_pad_with_physical_minimum(
-    physical_range: FloatRange,
+    physical_range: _FloatRange,
 ):
     test_data = np.arange(physical_range[0], physical_range[1] + 1)
     edf = Edf(
@@ -1457,13 +1159,3 @@ def test_sampling_frequencies_leading_to_floating_point_issues_in_signal_duratio
     assert edf.num_data_records == 10
     assert edf.signals[0].samples_per_data_record == 22
     assert edf.signals[1].samples_per_data_record == 9
-
-
-def test_edf_signal_init_does_not_accept_edf_annotations_as_label():
-    with pytest.raises(ValueError, match="must not be 'EDF Annotations'"):
-        EdfSignal(np.arange(2), 1, label="EDF Annotations")
-
-
-def test_edf_signal_label_cannot_be_set_to_edf_annotations():
-    with pytest.raises(ValueError, match="must not be 'EDF Annotations'"):
-        EdfSignal(np.arange(2), 1).label = "EDF Annotations"
