+++ tests/test_utils.py
@@ -1,5 +1,4 @@
 import datetime
-import math
 
 import pytest
 
@@ -8,7 +7,6 @@ from edfio._utils import (
     encode_annotation_duration,
     encode_annotation_onset,
     encode_edfplus_date,
-    round_float_to_8_characters,
 )
 
 VALID_EDFPLUS_DATE_PAIRS = (
@@ -82,45 +80,3 @@ def test_encode_annotation_duration(duration: float, expected: str):
 def test_encode_annotation_duration_raises_error_for_negative_values():
     with pytest.raises(ValueError, match="Annotation duration must be positive, is"):
         encode_annotation_duration(-1)
-
-
-# fmt: off
-@pytest.mark.parametrize(
-    (   "value",       "expected_round", "expected_floor", "expected_ceil"),
-    [
-        (1.1111114,     1.111111,         1.111111,         1.111112,),
-        (1.111111444,   1.111111,         1.111111,         1.111112,),
-        (1.111111499,   1.111111,         1.111111,         1.111112,),
-        (1.1111115,     1.111112,         1.111111,         1.111112,),
-        (1111111.4,     1111111,          1111111,          1111112,),
-        (1111111.444,   1111111,          1111111,          1111112,),
-        (1111111.499,   1111111,          1111111,          1111112,),
-        (1111111.5,     1111112,          1111111,          1111112,),
-        (11111111.4,    11111111,         11111111,         11111112,),
-        (11111111.444,  11111111,         11111111,         11111112,),
-        (11111111.499,  11111111,         11111111,         11111112,),
-        (11111111.5,    11111112,         11111111,         11111112,),
-        (-1.111114,     -1.11111,         -1.11112,         -1.11111,),
-        (-1.11111444,   -1.11111,         -1.11112,         -1.11111,),
-        (-1.11111499,   -1.11111,         -1.11112,         -1.11111,),
-        (-1.111115,     -1.11112,         -1.11112,         -1.11111,),
-        (-111111.4,     -111111,          -111112,          -111111,),
-        (-111111.444,   -111111,          -111112,          -111111,),
-        (-111111.499,   -111111,          -111112,          -111111,),
-        (-111111.5,     -111112,          -111112,          -111111,),
-        (-1111111.4,    -1111111,         -1111112,         -1111111,),
-        (-1111111.444,  -1111111,         -1111112,         -1111111,),
-        (-1111111.499,  -1111111,         -1111112,         -1111111,),
-        (-1111111.5,    -1111112,         -1111112,         -1111111,),
-    ],
-)
-# fmt: on
-def test_round_float_to_8_characters(
-    value: float,
-    expected_round: float,
-    expected_floor: float,
-    expected_ceil: float,
-):
-    assert round_float_to_8_characters(value, round) == expected_round
-    assert round_float_to_8_characters(value, math.floor) == expected_floor
-    assert round_float_to_8_characters(value, math.ceil) == expected_ceil
